chore: move spagateway needed functions to cryptoHelpers
